made responsive and changes file names
